requirements.txt
fastapi[all]
python-dotenv
groq